Set up authentications views
With superuser created two groups: Librarians and Library Members, also created two members and one librarian and added them to their groups. Created templates for user login, logout and password-reset operation, added an option to my views and basic template that checks whether the user is logged in or logged out and after that concludes what views can operate for the user.